@js00070 2019-10-18
> Why not use the same process control struct in non-vec function. Now it is too difficult to review and check the correctness.
> 
> ```
> forceToSigned := s.ctx.GetSessionVars().SQLMode.HasNoUnsignedSubtractionMode()
> isLHSUnsigned := !forceToSigned && mysql.HasUnsignedFlag(s.args[0].GetType().Flag)
> isRHSUnsigned := !forceToSigned && mysql.HasUnsignedFlag(s.args[1].GetType().Flag)
> 
> if forceToSigned && mysql.HasUnsignedFlag(s.args[0].GetType().Flag) {
> 	for i := 0;i < n;i++ {
> 		if lhi64s[i] < 0 || (lhi64s[i] > math.MaxInt64) {
> 			return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", s.args[0].String(), s.args[1].String()))
> 		}
> 	}
> }
> if forceToSigned && mysql.HasUnsignedFlag(s.args[1].GetType().Flag) {
> ...
> }
> switch {
> ...
> }
> for i := 0;i < n;i++ {
> resulti64s[i] = lhi64s[i]-rhi64i[i]
> }
> return nil
> ```
> 
> I think this struct is good. Consider it?

I take a try and find a potential problem in this code structrue.

in the non-vec function, if `args[0]` or `args[1]` is `Null`, the function will return `IsNull` immediately and will not check the following `BIGINT` condition.

but in vectorized version, if we use this code structure, we have to add `if buf.IsNull(i)` in every `for` loop of all check step to get the exactly the same process of non-vec version. If we don't add, the behavior may be different in some cases? but adding so many `IsNull` may cause performance degradation in large amount of rows.