@js00070 2019-10-19
> > > Why not use the same process control struct in non-vec function. Now it is too difficult to review and check the correctness.
> > > ```
> > > forceToSigned := s.ctx.GetSessionVars().SQLMode.HasNoUnsignedSubtractionMode()
> > > isLHSUnsigned := !forceToSigned && mysql.HasUnsignedFlag(s.args[0].GetType().Flag)
> > > isRHSUnsigned := !forceToSigned && mysql.HasUnsignedFlag(s.args[1].GetType().Flag)
> > > 
> > > if forceToSigned && mysql.HasUnsignedFlag(s.args[0].GetType().Flag) {
> > > 	for i := 0;i < n;i++ {
> > > 		if lhi64s[i] < 0 || (lhi64s[i] > math.MaxInt64) {
> > > 			return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", s.args[0].String(), s.args[1].String()))
> > > 		}
> > > 	}
> > > }
> > > if forceToSigned && mysql.HasUnsignedFlag(s.args[1].GetType().Flag) {
> > > ...
> > > }
> > > switch {
> > > ...
> > > }
> > > for i := 0;i < n;i++ {
> > > resulti64s[i] = lhi64s[i]-rhi64i[i]
> > > }
> > > return nil
> > > ```
> > > 
> > > 
> > > I think this struct is good. Consider it?
> > 
> > 
> > I take a try and find a potential problem in this code structrue.
> > in the non-vec function, if `args[0]` or `args[1]` is `Null`, the function will return `IsNull` immediately and will not check the following `BIGINT` condition.
> > but in vectorized version, if we use this code structure, we have to add `if buf.IsNull(i)` in every `for` loop of all check step to get the exactly the same process of non-vec version. If we don't add, the behavior may be different in some cases? but adding so many `IsNull` may cause performance degradation in large amount of rows.
> 
> How about introduce a int64 slice to record the index which row is null?
> 
> ```
> MergeNull
> indexes := make([]int64,n)
> for i := 0;i < n;i++{
>     if !isnull(i) indexes = append(indexes,i)
> }
> 
> if check {
>     for _, idx := range indexes {
>         if check {
>             return err   
>         }
>     }
> }
> ```
> 
> I don't know how much the performance degradation in so many `IsNull`, we can have a test.
> If the result is not really good, maybe we need abandon the old struct and try a new struct(like you code now).

I benchtested all the code versions we mentioned and got the compare results.

I runed the benchtest using the following commands
```
go test -v -benchmem -bench=BenchmarkVectorizedBuiltinArithmeticFunc -count=20 -run=BenchmarkVectorizedBuiltinArithmeticFunc | tee mycode.txt

go test -v -benchmem -bench=BenchmarkVectorizedBuiltinArithmeticFunc -count=20 -run=BenchmarkVectorizedBuiltinArithmeticFunc | tee new1.txt

go test -v -benchmem -bench=BenchmarkVectorizedBuiltinArithmeticFunc -count=20 -run=BenchmarkVectorizedBuiltinArithmeticFunc | tee new2.txt
```

and the result is
```bash
benchstat mycode.txt new1.txt
name                                                                               old time/op    new time/op    delta
VectorizedBuiltinArithmeticFunc/builtinArithmeticMinusIntSig-VecBuiltinFunc-12       4.88µs ± 1%    6.91µs ± 1%  +41.76%  (p=0.000 n=20+20)

benchstat mycode.txt new2.txt
name                                                                               old time/op    new time/op     delta
VectorizedBuiltinArithmeticFunc/builtinArithmeticMinusIntSig-VecBuiltinFunc-12       4.88µs ± 1%    36.92µs ± 2%  +657.11%  (p=0.000 n=20+20)

```

the `mycode.txt` is the benchtest result of my original code

the `new1.txt` is the benchtest result of the code which uses tons of `IsNull(i)`, code as follow:
```go
	forceToSigned := b.ctx.GetSessionVars().SQLMode.HasNoUnsignedSubtractionMode()
	isLHSUnsigned := !forceToSigned && mysql.HasUnsignedFlag(b.args[0].GetType().Flag)
	isRHSUnsigned := !forceToSigned && mysql.HasUnsignedFlag(b.args[1].GetType().Flag)

	if forceToSigned && mysql.HasUnsignedFlag(b.args[0].GetType().Flag) {
		for i := 0; i < n; i++ {
			if result.IsNull(i) {
				continue
			}
			if lhi64s[i] < 0 || (lhi64s[i] > math.MaxInt64) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
		}
	}

	if forceToSigned && mysql.HasUnsignedFlag(b.args[1].GetType().Flag) {
		for i := 0; i < n; i++ {
			if result.IsNull(i) {
				continue
			}
			if rhi64s[i] < 0 || (rhi64s[i] > math.MaxInt64) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
		}
	}

	switch {
	case isLHSUnsigned && isRHSUnsigned:
		for i := 0; i < n; i++ {
			if result.IsNull(i) {
				continue
			}
			if uint64(lhi64s[i]) < uint64(rhi64s[i]) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
		}
	case isLHSUnsigned && !isRHSUnsigned:
		for i := 0; i < n; i++ {
			if result.IsNull(i) {
				continue
			}
			if rhi64s[i] >= 0 && uint64(lhi64s[i]) < uint64(rhi64s[i]) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
			if rhi64s[i] < 0 && uint64(lhi64s[i]) > math.MaxUint64-uint64(-rhi64s[i]) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
		}
	case !isLHSUnsigned && isRHSUnsigned:
		for i := 0; i < n; i++ {
			if result.IsNull(i) {
				continue
			}
			if uint64(lhi64s[i]-math.MinInt64) < uint64(rhi64s[i]) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
		}
	case !isLHSUnsigned && !isRHSUnsigned:
		for i := 0; i < n; i++ {
			if result.IsNull(i) {
				continue
			}
			if (lhi64s[i] > 0 && -rhi64s[i] > math.MaxInt64-lhi64s[i]) || (lhi64s[i] < 0 && -rhi64s[i] < math.MinInt64-lhi64s[i]) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
		}
	}
	for i := 0; i < n; i++ {
		if result.IsNull(i) {
			continue
		}
		resulti64s[i] = lhi64s[i] - rhi64s[i]
	}

	return err
```

the `new2.txt` is the benchtest result of the code which uses a `indexes` slice to replace `IsNull` just like you mentioned, code as follow:

```go
	indexes := make([]int, n)
	for i := 0; i < n; i++ {
		if !result.IsNull(i) {
			indexes = append(indexes, int(i))
		}
	}

	if forceToSigned && mysql.HasUnsignedFlag(b.args[0].GetType().Flag) {
		for _, i := range indexes {
			if lhi64s[i] < 0 || (lhi64s[i] > math.MaxInt64) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
		}
	}

	if forceToSigned && mysql.HasUnsignedFlag(b.args[1].GetType().Flag) {
		for _, i := range indexes {
			if rhi64s[i] < 0 || (rhi64s[i] > math.MaxInt64) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
		}
	}

	switch {
	case isLHSUnsigned && isRHSUnsigned:
		for _, i := range indexes {
			if uint64(lhi64s[i]) < uint64(rhi64s[i]) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
		}
	case isLHSUnsigned && !isRHSUnsigned:
		for _, i := range indexes {
			if rhi64s[i] >= 0 && uint64(lhi64s[i]) < uint64(rhi64s[i]) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
			if rhi64s[i] < 0 && uint64(lhi64s[i]) > math.MaxUint64-uint64(-rhi64s[i]) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
		}
	case !isLHSUnsigned && isRHSUnsigned:
		for _, i := range indexes {
			if uint64(lhi64s[i]-math.MinInt64) < uint64(rhi64s[i]) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT UNSIGNED", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
		}
	case !isLHSUnsigned && !isRHSUnsigned:
		for _, i := range indexes {
			if (lhi64s[i] > 0 && -rhi64s[i] > math.MaxInt64-lhi64s[i]) || (lhi64s[i] < 0 && -rhi64s[i] < math.MinInt64-lhi64s[i]) {
				return types.ErrOverflow.GenWithStackByArgs("BIGINT", fmt.Sprintf("(%s - %s)", b.args[0].String(), b.args[1].String()))
			}
		}
	}
	for _, i := range indexes {
		resulti64s[i] = lhi64s[i] - rhi64s[i]
	}
```

We can see my origin code is the fastest. I think we should keep it.

BTW, it's pretty werid that the `new2` version is so much slower than the `new1` version. Can anyone tell the reason? I guess it may be about the CPU hardware optimization but not really sure.